src/regfmt/requirements.txt
attrs==22.1.0
jsonschema==4.16.0
Pillow==9.2.0
pyrsistent==0.18.1
PyYAML==6.0
tinycss2==1.1.1
webencodings==0.5.1
